requirements.txt
tweepy>=4.14.0
feedparser>=6.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
groq>=0.4.0
requests>=2.31.0
pytz>=2024.1
